requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
httpx>=0.27
pydantic>=2.6
requests>=2.31
openai>=1.0
pillow
torchvision
